removed the file that was not implemented yet. Also split the self-calibration into a separate file
